Revise email script to handle dates better.
Previously, we stuck the date from the patch into the email headers.
Now we stick that in the email body, and stick the _merge_ date into
the header.  This way emails show up in the order that stuff is merged
into the repo, rather than in the order that the patches were
originally written.